Indexer.go broken into three files. All cases seperated into seperate funtions. Started log15 transiton.
